// generated/models/ApiWalletDistributionAllocation.ts
// @ts-nocheck
/**
 * 6529.io API
 * JSON REST API for 6529.io. New here? A terminology overview and a step-by-step authentication walkthrough (guides &amp; auth) live at <a href=\"https://6529.io/tools/api\">https://6529.io/tools/api</a>. The raw machine-readable spec is downloadable at <a href=\"/openapi.yaml\">/openapi.yaml</a> and <a href=\"/openapi.json\">/openapi.json</a>.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class ApiWalletDistributionAllocation {
    'phase': ApiWalletDistributionAllocationPhaseEnum;
    'spots_airdrop': number;
    'spots_allowlist': number;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "phase",
            "baseName": "phase",
            "type": "ApiWalletDistributionAllocationPhaseEnum",
            "format": ""
        },
        {
            "name": "spots_airdrop",
            "baseName": "spots_airdrop",
            "type": "number",
            "format": "int64"
        },
        {
            "name": "spots_allowlist",
            "baseName": "spots_allowlist",
            "type": "number",
            "format": "int64"
        }    ];

    static getAttributeTypeMap() {
        return ApiWalletDistributionAllocation.attributeTypeMap;
    }

    public constructor() {
    }
}

export enum ApiWalletDistributionAllocationPhaseEnum {
    Phase0 = 'Phase 0',
    Phase1 = 'Phase 1',
    Phase2 = 'Phase 2',
    Public = 'Public'
}